Atualização e otimização da feature
Otimização dos sistemas de geração e edição, (os sistemas foram separados em classes independentes)

+++ TrimSheet_Tools/Model/Vector2D.cs
@@ -9,12 +9,12 @@ namespace TrimSheet_Tools.Model
     public class Vector2D
     {
         // Global variables
-        public float X { get; set; }
-        public float Y { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
 
 
         // Constructor
-        public Vector2D(float x, float y)
+        public Vector2D(double x, double y)
         { X = x; Y = y; }
         public Vector2D() : this(0, 0) { }
 
@@ -41,12 +41,12 @@ namespace TrimSheet_Tools.Model
         public static Vector2D operator -(Vector2D a, Vector2D b)
         { return new Vector2D(a.X - b.X, a.Y - b.Y); }
 
-        public static Vector2D operator *(Vector2D vector, float scalar)
+        public static Vector2D operator *(Vector2D vector, double scalar)
         { return new Vector2D(vector.X * scalar, vector.Y * scalar); }
-        public static Vector2D operator *(float scalar, Vector2D vector)
+        public static Vector2D operator *(double scalar, Vector2D vector)
         { return vector * scalar; }
 
-        public static Vector2D operator /(Vector2D vector, float scalar)
+        public static Vector2D operator /(Vector2D vector, double scalar)
         {
             if (Math.Abs(scalar) < float.Epsilon)
                 throw new DivideByZeroException("You cannot divide a vector by zero.");
